revert qcp/ -> src/ rename
