+++ tox.ini
@@ -14,17 +14,17 @@ usedevelop = False
 commands = flake8
 
 [testenv:tests]
-commands = pytest --cache-clear --verbose --cov=src \
+commands = pytest --cache-clear --verbose --cov=qcp \
               --cov-report=xml:coverage/python/coverage.xml \
               --cov-report=term \
               --junitxml=test-reports/pytest/results.xml {posargs} \
-              {toxinidir}/src/tests
+              {toxinidir}/qcp/tests
 setenv =
        RUNNING_IN_TOX = 1
 
 
 [testenv:mypy]
-commands = mypy -p src
+commands = mypy -p qcp
 
 [testenv:update-pins]
 envdir = {toxworkdir}/pip_tools
